+++ lib/src/bparts.dart
@@ -14,8 +14,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import 'package:fmatch/fmatch.dart';
-
 import '../fmscreen.dart';
 import 'util.dart';